Investigation:
It's possible to make JPT_ENUM local scoped, with the trade off where constexpr and templated parameters won't be available anymore.

At this point, I think performance and programming efficiency are prioritized as local scoped enum could be replaced with raw built-in enum class, and JPT_ENUM seems to be needed globally.

In future, if we have to modify JPT_ENUM to be local scoped, Four things to modify:

1. Remove "inline static" from const EnumData<TSize> s_data = GenerateData<TSize>(#__VA_ARGS__);

2. Remove all the constexpr :(

3. Modify the static functions like TSize Min(), Max(), Count(), etc. Follow this pattern
static TSize Min()
{
    EnumName e;
    return e.s_data.min;
}

4. Remove all the "template<jpt::Integral TInt = TSize>"